requirements.txt
discord.ext.context
discord.py